// projects/ng-wizard/src/lib/utils/enums.ts
export enum TOOLBAR_POSITION {
    none = 'none',
    top = 'top',
    bottom = 'bottom',
    both = 'both'
}

export enum TOOLBAR_BUTTON_POSITION {
    start = 'start',
    end = 'end'
}

// export enum TRANSITION_EFFECT {
//     none = 'none',
//     slide = 'slide',
//     fade = 'fade'
// }

export enum THEME {
    default = 'default',
    arrows = 'arrows',
    circles = 'circles',
    dots = 'dots'
}

export enum STEP_STATE {
    normal = 'normal',
    disabled = 'disabled',
    error = 'error',
    hidden = 'hidden'
}

export enum STEP_STATUS {
    untouched = 'untouched',
    done = 'done',
    active = 'active'
}

export enum STEP_DIRECTION {
    forward = 'forward',
    backward = 'backward'
}

export enum STEP_POSITION {
    first = 'first',
    final = 'final',
    middle = 'middle'
}